Change behavior to return all matches and as a result be thread safe.
Return from Match now also returns position as well as dictionary index.
This has a knock on effect we must return all matches, not just the
first. An added bonus from this is that the Match call is now thread
safe, as it doesn't mutate the data structure whilst running.